Detect delimiter for #159
The Python csv.Sniffer() class failed, because it chokes on accented
Unicode characters.

Instead, start with a naive implementation that just counts possible
in a sample and chooses the one that appears the most.

TODO: we should also consider per-line distribution. A tab-separated
file with a lot of text could have many embedded commas.