Fix crash when reading a local file in a production environment (#1)
When using this code in an environment where the python libraries are in a virtualenv where the files are readable but not writable will crash due to a change that was made to use a permissions string of 'rb+'. This file shouldn't be updated by this reader code (I think), so revert it back to 'rb'.